+++ core-ng-test/src/test/java/core/framework/test/TestModule.java
@@ -55,6 +55,7 @@ public class TestModule extends AbstractTestModule {
 
     private void configureAPI() {
         api().client(TestWebService.class, "https://localhost:8443").intercept(new TestWebServiceClientInterceptor());
+        api().publishAPI(List.of("0.0.0.0/0"));
     }
 
     private void configureRedis() {
@@ -81,7 +82,6 @@ public class TestModule extends AbstractTestModule {
         site().session().cookie("SessionId", "localhost");
         site().cdn().host("//cdn");
         site().security().contentSecurityPolicy("default-src 'self' https://cdn; img-src 'self' https://cdn data:; object-src 'none'; frame-src 'none';");
-        site().publishAPI(List.of("0.0.0.0/0"));
     }
 
     private void configureHTTP() {